+++ market_economy_simulation/src/terrain_storage.rs
@@ -1,392 +1,240 @@
 //! Manages the data on the gpu of the terrain
 //!
 
-pub mod lod_heightmap_tile;
 pub mod lod_quad_tree;
-pub mod terrain_storage_v2;
 pub mod quad_tree_draw;
 pub mod terrain_texture_details;
 
-// use std::sync::mpsc;
+use std::sync::mpsc;
 
-// use cgmath::{MetricSpace, Zero};
-// use market_economy_simulation_server::{game_logic::game_logic_interface::GameLogicMessageRequest, heightmap_generator};
-// use wgpu_renderer::{
-//     shape::{self, MeshDataInterface},
-//     vertex_heightmap_shader,
-//     wgpu_renderer::WgpuRendererInterface,
-// };
+use cgmath::Zero;
+use lod_quad_tree::LodQuadTree;
+use market_economy_simulation_server::{
+    game_logic::game_logic_interface::GameLogicMessageRequest, heightmap_generator,
+};
+use quad_tree_draw::QuadTreeDraw;
+use terrain_texture_details::TerrainTextureDetails;
+use wgpu_renderer::{
+    shape::{self, MeshDataInterface},
+    vertex_heightmap_shader,
+    wgpu_renderer::WgpuRendererInterface,
+};
 
-// use crate::deferred_heightmap_shader::{self, DeferredHeightMapShaderDraw};
+use crate::deferred_heightmap_shader::{self, DeferredHeightMapShaderDraw};
 
-// const LOD: usize = 4;
+pub struct TerrainSettings {
+    pub nr_tiles: usize,
+    pub max_depth: usize,
+}
 
-// type HeightmapInstanceBuffer =
-//     deferred_heightmap_shader::InstanceBuffer<deferred_heightmap_shader::Instance>;
+pub struct TerrainStorage {
+    settings: TerrainSettings,
 
-// pub struct TerrainSettings {
-//     pub terrain_tile_size: usize,
-//     pub terrain_size: (usize, usize),
-// }
+    mesh: deferred_heightmap_shader::Mesh,
+    texture: deferred_heightmap_shader::Texture,
+    heightmap_textures: Vec<deferred_heightmap_shader::HeightmapTexture>,
+    instances: Vec<deferred_heightmap_shader::InstanceBuffer<deferred_heightmap_shader::Instance>>,
+    heightmap_details: Vec<TerrainTextureDetails>,
 
-// pub struct TerrainStorage {
-//     settings: TerrainSettings,
+    lod_quad_tree: lod_quad_tree::LodQuadTree,
 
-//     mesh: [deferred_heightmap_shader::Mesh; LOD],
-//     texture: deferred_heightmap_shader::Texture,
+    view_position: cgmath::Vector3<isize>,
 
-//     instances: Vec<[HeightmapInstanceBuffer; LOD]>,
-//     height_textures: Vec<[Option<deferred_heightmap_shader::HeightmapTexture>; LOD]>,
-//     height_texture_details: Vec<[heightmap_generator::HeightMapDetails; LOD]>,
-//     height_textures_state: Vec<[HeightTextureState; LOD]>,
-//     texture_positions: Vec<cgmath::Vector3<f32>>,
+    max_depth: usize,
+    nr_tiles: usize,
 
-//     view_position: cgmath::Vector3<f32>,
+    requests: Vec<TerrainTextureDetails>,
+}
 
-//     width: usize,
-//     height: usize,
-//     size: usize,
-//     tile_size: usize,
+impl TerrainStorage {
+    pub fn new(
+        settings: TerrainSettings,
+        renderer: &mut dyn WgpuRendererInterface,
+        texture_bind_group_layout: &deferred_heightmap_shader::TextureBindGroupLayout,
+        // heightmap_bind_group_layout: &deferred_heightmap_shader::HeightmapBindGroupLayout,
+    ) -> Self {
+        let max_depth = settings.max_depth;
+        let nr_tiles = settings.nr_tiles;
+        let size_0 = nr_tiles + 3;
+        let size_1 = nr_tiles + 1;
 
-//     requests: Vec<heightmap_generator::HeightMapDetails>,
-// }
+        // mesh
+        let grid = shape::Grid::new(1.0, size_1, 1);
+        let gird_triangles = grid.triangles();
+        let mesh = deferred_heightmap_shader::Mesh::from_shape(renderer.device(), gird_triangles);
 
-// impl TerrainStorage {
-//     pub fn new(
-//         settings: TerrainSettings,
-//         renderer: &mut dyn WgpuRendererInterface,
-//         texture_bind_group_layout: &deferred_heightmap_shader::TextureBindGroupLayout,
-//         heightmap_bind_group_layout: &deferred_heightmap_shader::HeightmapBindGroupLayout,
-//     ) -> Self {
-//         let width = settings.terrain_size.0;
-//         let height = settings.terrain_size.1;
-//         let size: usize = width * height;
-//         let texture_side_length = settings.terrain_tile_size + 1;
+        // texture
+        let texture_bytes = include_bytes!("../res/tile.png");
+        let texture_image = image::load_from_memory(texture_bytes).unwrap();
+        let texture_rgba = texture_image.to_rgba8();
+        let texture = deferred_heightmap_shader::Texture::new_with_mipmaps(
+            renderer,
+            &texture_bind_group_layout,
+            &texture_rgba,
+            Some("tile.png"),
+            9,
+        )
+        .unwrap();
 
-//         // mesh
-//         let mesh: [deferred_heightmap_shader::Mesh; LOD] = std::array::from_fn(|index| {
-//             let a = 2u32.pow(index as u32);
+        // heightmap_textures
+        let heightmap_textures = Vec::new();
+        let instances = Vec::new();
+        let heightmap_details = Vec::new();
 
-//             let grid = shape::Grid::new(1.0, settings.terrain_tile_size / (a as usize) + 1);
-//             let gird_triangles = grid.triangles();
-//             deferred_heightmap_shader::Mesh::from_shape(renderer.device(), gird_triangles)
-//         });
+        // lod_quad_tree
+        let lod_quad_tree = LodQuadTree::new(max_depth, nr_tiles);
 
-//         // texture
-//         let texture_bytes = include_bytes!("../res/tile.png");
-//         // let texture_bytes = include_bytes!("../res/pony2.png");
-//         let texture_image = image::load_from_memory(texture_bytes).unwrap();
-//         let texture_rgba = texture_image.to_rgba8();
+        // view_position
+        let view_position = cgmath::Vector3::zero();
 
-//         let texture = deferred_heightmap_shader::Texture::new_with_mipmaps(
-//             renderer,
-//             &texture_bind_group_layout,
-//             &texture_rgba,
-//             Some("tile.png"),
-//             9,
-//         )
-//         .unwrap();
+        // requests
+        let requests = Vec::new();
 
-//         // position
-//         let offset_y = (texture_side_length - 1) * height / 2;
-//         let offset_x = (texture_side_length - 1) * width / 2;
-//         let mut texture_positions: Vec<cgmath::Vector3<f32>> = Vec::with_capacity(size);
-//         for y in 0..height {
-//             for x in 0..width {
-//                 let pos = cgmath::Vector3::new(
-//                     (x * (texture_side_length - 1)) as f32 - offset_x as f32,
-//                     (y * (texture_side_length - 1)) as f32 - offset_y as f32,
-//                     0.0,
-//                 );
-//                 texture_positions.push(pos);
-//             }
-//         }
+        Self {
+            settings,
+            mesh,
+            texture,
+            heightmap_textures,
+            instances,
+            heightmap_details,
+            lod_quad_tree,
+            view_position,
+            max_depth,
+            nr_tiles,
+            requests,
+        }
+    }
 
-//         let mut instances: Vec<[HeightmapInstanceBuffer; LOD]> = Vec::with_capacity(size);
-//         for i in 0..size {
-//             let instance_array: [HeightmapInstanceBuffer; LOD] = std::array::from_fn(|index| {
-//                 let a = 2u32.pow(index as u32);
+    pub fn update_height_map(
+        &mut self,
+        renderer: &mut dyn WgpuRendererInterface,
+        heightmap_bind_group_layout: &deferred_heightmap_shader::HeightmapBindGroupLayout,
+        height_map: heightmap_generator::HeightMap,
+    ) {
+        let pos_0 = height_map.details.pos_0;
+        let pos_1 = height_map.details.pos_1;
+        let point_distance = height_map.details.point_distance;
 
-//                 let instance = deferred_heightmap_shader::Instance {
-//                     position: texture_positions[i].into(),
-//                     color: cgmath::Vector3::new(0.2, 0.2, 0.8).into(),
-//                     entity: i as u32,
-//                     distance: a as f32,
-//                 };
-//                 HeightmapInstanceBuffer::new(renderer.device(), &[instance])
-//             });
+        let size_0 = height_map.details.size_0;
+        let size_1 = height_map.details.size_1;
 
-//             instances.push(instance_array);
-//         }
+        let nr_tiles = height_map.details.nr_tiles;
 
-//         let mut height_textures: Vec<[Option<deferred_heightmap_shader::HeightmapTexture>; LOD]> =
-//             Vec::with_capacity(size);
-//         for i in 0..size {
-//             let height_texture_array: [Option<deferred_heightmap_shader::HeightmapTexture>; LOD] =
-//                 std::array::from_fn(|index| None);
-//             height_textures.push(height_texture_array);
-//         }
+        let depth = height_map.details.depth;
+        let node_index = height_map.details.node_index;
 
-//         let mut height_texture_details: Vec<[heightmap_generator::HeightMapDetails; LOD]> =
-//             Vec::with_capacity(size);
-//         for y in 0..height {
-//             for x in 0..width {
-//                 let height_texture_details_array: [heightmap_generator::HeightMapDetails; LOD] =
-//                     std::array::from_fn(|lod| {
-//                         let a = 2u32.pow(lod as u32);
+        assert_eq!(nr_tiles, self.nr_tiles);
 
-//                         heightmap_generator::HeightMapDetails {
-//                             distance: a as usize,
-//                             size_x: texture_side_length / a as usize,
-//                             size_y: texture_side_length / a as usize,
-//                             x: x * (texture_side_length - 1),
-//                             y: y * (texture_side_length - 1),
-//                             index: y * width + x,
-//                             lod,
-//                         }
-//                     });
-//                     height_texture_details.push(height_texture_details_array);
-//             }
-//         }
+        // create host data
+        let mut heightmap: Vec<deferred_heightmap_shader::Heightmap> =
+            Vec::with_capacity(size_0 * size_0);
+        assert_eq!(height_map.heights.len(), size_0 * size_0);
+        for elem in height_map.heights {
+            heightmap.push(vertex_heightmap_shader::Heightmap { height: elem });
+            // heightmap.push(vertex_heightmap_shader::Heightmap { height: 0.0 });
+        }
 
-//         let mut height_textures_state: Vec<[HeightTextureState; LOD]> = Vec::with_capacity(size);
-//         for i in 0..size {
-//             let height_textures_state_array: [HeightTextureState; LOD] =
-//                 std::array::from_fn(|index| HeightTextureState::NotRequested);
-//             height_textures_state.push(height_textures_state_array);
-//         }
+        let heightmap_details = TerrainTextureDetails {
+            pos_0,
+            pos_1,
+            point_distance,
+            size_0,
+            size_1,
+            nr_tiles,
+            depth,
+            node_index,
+        };
 
-//         // insert zero value at level 0
-//         // for i in 0..size {
-//         //     {
-//         //         let mut heightmap: Vec<deferred_heightmap_shader::Heightmap> =
-//         //             Vec::with_capacity(texture_side_length * texture_side_length);
-//         //         heightmap.resize(
-//         //             texture_side_length * texture_side_length,
-//         //             deferred_heightmap_shader::Heightmap::zero(),
-//         //         );
+        // create device data
+        let height_texture = deferred_heightmap_shader::HeightmapTexture::new(
+            renderer,
+            heightmap_bind_group_layout,
+            &heightmap,
+            size_0 as u32,
+            size_0 as u32,
+            Some("terrain"),
+        );
 
-//         //         let height_texture = deferred_heightmap_shader::HeightmapTexture::new(
-//         //             renderer.device(),
-//         //             heightmap_bind_group_layout,
-//         //             &heightmap,
-//         //             width as u32,
-//         //             height as u32,
-//         //             Some("terrain"),
-//         //         );
+        let instance = deferred_heightmap_shader::Instance {
+            position: [
+                (pos_1.x - point_distance as isize) as f32,
+                (pos_1.y - point_distance as isize) as f32,
+                0.0,
+            ],
+            color: [0.2, 0.2, 0.8],
+            entity: node_index as u32,
+            distance: point_distance as f32,
+        };
+        let instance_buffer =
+            deferred_heightmap_shader::InstanceBuffer::new(renderer.device(), &[instance]);
 
-//         //         height_textures[i][0] = Some(height_texture);
-//         //     }
-//         //     {
-//         //         let side_length = texture_side_length / 2;
-//         //         let mut heightmap: Vec<deferred_heightmap_shader::Heightmap> =
-//         //             Vec::with_capacity(side_length * side_length);
-//         //         heightmap.resize(
-//         //             side_length * side_length,
-//         //             deferred_heightmap_shader::Heightmap::zero(),
-//         //         );
+        // save device data
+        let data_index = self.heightmap_textures.len();
+        assert_eq!(self.instances.len(), data_index);
+        assert_eq!(self.heightmap_details.len(), data_index);
 
-//         //         let height_texture = deferred_heightmap_shader::HeightmapTexture::new(
-//         //             renderer.device(),
-//         //             heightmap_bind_group_layout,
-//         //             &heightmap,
-//         //             width as u32,
-//         //             height as u32,
-//         //             Some("terrain"),
-//         //         );
+        self.heightmap_textures.push(height_texture);
+        self.instances.push(instance_buffer);
+        self.heightmap_details.push(heightmap_details);
 
-//         //         height_textures[i][1] = Some(height_texture);
-//         //     }
+        // make node data available
+        self.lod_quad_tree.set_data_index(node_index, data_index);
+    }
 
-//         //     {
-//         //         let side_length = texture_side_length / 4;
-//         //         let mut heightmap: Vec<deferred_heightmap_shader::Heightmap> =
-//         //             Vec::with_capacity(side_length * side_length);
-//         //         heightmap.resize(
-//         //             side_length * side_length,
-//         //             deferred_heightmap_shader::Heightmap::zero(),
-//         //         );
+    pub fn update_view_position(&mut self, view_position: &cgmath::Vector3<f32>) {
+        self.view_position = cgmath::Vector3::new(
+            view_position.x as isize,
+            view_position.y as isize,
+            view_position.z as isize,
+        );
+    }
 
-//         //         let height_texture = deferred_heightmap_shader::HeightmapTexture::new(
-//         //             renderer.device(),
-//         //             heightmap_bind_group_layout,
-//         //             &heightmap,
-//         //             width as u32,
-//         //             height as u32,
-//         //             Some("terrain"),
-//         //         );
+    pub fn submit_requests(&mut self, sender: &mpsc::Sender<GameLogicMessageRequest>) {
+        for elem in &self.requests {
+            let _res = sender.send(GameLogicMessageRequest::GetTerrain(
+                heightmap_generator::HeightMapDetails {
+                    pos_0: elem.pos_0,
+                    pos_1: elem.pos_1,
+                    point_distance: elem.point_distance,
+                    size_0: elem.size_0,
+                    size_1: elem.size_1,
+                    nr_tiles: elem.nr_tiles,
+                    depth: elem.depth,
+                    node_index: elem.node_index,
+                },
+            ));
+        }
 
-//         //         height_textures[i][2] = Some(height_texture);
-//         //     }
+        self.requests.clear();
+    }
+}
 
-//         //     {
-//         //         let side_length = texture_side_length / 8;
-//         //         let mut heightmap: Vec<deferred_heightmap_shader::Heightmap> =
-//         //             Vec::with_capacity(side_length * side_length);
-//         //         heightmap.resize(
-//         //             side_length * side_length,
-//         //             deferred_heightmap_shader::Heightmap::zero(),
-//         //         );
+impl DeferredHeightMapShaderDraw for TerrainStorage {
+    fn draw<'a>(&'a mut self, render_pass: &mut wgpu::RenderPass<'a>) {
+        // mesh data
+        let mesh = &self.mesh;
+        let texture = &self.texture;
+        let heightmap_textures = &self.heightmap_textures;
+        let instances = &self.instances;
 
-//         //         let height_texture = deferred_heightmap_shader::HeightmapTexture::new(
-//         //             renderer.device(),
-//         //             heightmap_bind_group_layout,
-//         //             &heightmap,
-//         //             width as u32,
-//         //             height as u32,
-//         //             Some("terrain"),
-//         //         );
+        // quad tree data
+        let lod_quad_tree = &mut self.lod_quad_tree;
+        let view_position = &self.view_position;
+        let max_depth = self.max_depth;
+        let nr_tiles = self.nr_tiles;
 
-//         //         height_textures[i][3] = Some(height_texture);
-//         //     }
-//         // }
+        // draw
+        mesh.bind(render_pass);
+        texture.bind(render_pass);
 
-//         let view_position: cgmath::Vector3<f32> = cgmath::Vector3::zero();
+        let quad_tree_draw = &mut QuadTreeDraw::new(max_depth, nr_tiles, |data_index| {
+            heightmap_textures[data_index].bind(render_pass);
+            instances[data_index].bind(render_pass);
+            mesh.draw_indexed(render_pass);
+        });
 
-//         let tile_size = settings.terrain_tile_size;
+        lod_quad_tree.traverse_leaves(view_position, quad_tree_draw);
 
-//         Self {
-//             settings,
-//             mesh,
-//             texture,
-//             instances,
-//             height_textures,
-//             height_texture_details,
-//             height_textures_state,
-//             texture_positions,
-//             view_position,
-//             size,
-//             width,
-//             height,
-//             tile_size,
-//             requests: Vec::new(),
-//         }
-//     }
-
-//     fn calculate_lod_index(
-//         view_position: cgmath::Vector3<f32>,
-//         position: cgmath::Vector3<f32>,
-//         tile_size: usize,
-//     ) -> usize {
-//         let distance = view_position.distance(position);
-//         // if distance > tile_size as f32 * 6.0 {
-//         //     3
-//         // } else if distance > tile_size as f32 * 4.0 {
-//         //     2
-//         // } else if distance > tile_size as f32 * 2.0 {
-//         //     1
-//         // } else {
-//         //     0
-//         // }
-
-//         2
-//     }
-
-//     pub fn update_view_position(&mut self, view_position: &cgmath::Vector3<f32>) {
-//         self.view_position = *view_position;
-//     }
-
-//     pub fn submit_requests(&mut self, sender: &mpsc::Sender<GameLogicMessageRequest>) {
-//         for elem in &self.requests {
-//             sender.send(GameLogicMessageRequest::GetTerrain(elem.clone()));
-//         }
-
-//         self.requests.clear();
-//     }
-
-//     pub(crate) fn update_height_map(
-//         &mut self,
-//         renderer: &mut dyn WgpuRendererInterface,
-//         heightmap_bind_group_layout: &deferred_heightmap_shader::HeightmapBindGroupLayout,
-//         height_map: market_economy_simulation_server::heightmap_generator::HeightMap,
-//     ) {
-//         let distance = height_map.details.distance;
-//         let size_x = height_map.details.size_x;
-//         let size_y = height_map.details.size_y;
-//         let p_x = height_map.details.x;
-//         let p_y = height_map.details.y;
-
-//         let size = size_x * size_y;
-
-//         // create host data
-//         let mut heightmap: Vec<deferred_heightmap_shader::Heightmap> = Vec::with_capacity(size);
-
-//         assert_eq!(height_map.heights.len(), size);
-
-//         for elem in height_map.heights {
-//             heightmap.push(vertex_heightmap_shader::Heightmap { height: elem });
-//         }
-
-//         // create device data
-//         let height_texture = deferred_heightmap_shader::HeightmapTexture::new(
-//             renderer,
-//             heightmap_bind_group_layout,
-//             &heightmap,
-//             size_x as u32,
-//             size_y as u32,
-//             Some("terrain"),
-//         );
-
-//         // set result
-//         let index = height_map.details.index;
-//         let lod = height_map.details.lod;
-//         self.height_textures[index][lod] = Some(height_texture);
-//         self.height_textures_state[index][lod] = HeightTextureState::Available;
-//     }
-// }
-
-// impl DeferredHeightMapShaderDraw for TerrainStorage {
-//     fn draw<'a>(&'a mut self, render_pass: &mut wgpu::RenderPass<'a>) {
-//         for i in 0..self.size {
-//             // get data at index
-//             let position = &self.texture_positions[i];
-
-//             // level of detail
-//             let lod = Self::calculate_lod_index(self.view_position, *position, self.tile_size);
-//             let mesh = &self.mesh[lod];
-//             let texture = &self.texture;
-//             let height_texture = &self.height_textures[i][lod];
-//             let instance = &self.instances[i][lod];
-
-//             // draw
-//             match self.height_textures_state[i][lod] {
-//                 HeightTextureState::NotRequested => {
-//                     self.requests.push(self.height_texture_details[i][lod].clone());
-//                     self.height_textures_state[i][lod] = HeightTextureState::IsRequested;
-//                 }
-//                 HeightTextureState::IsRequested => {
-//                     // nothing to do, waiting for a result
-//                 }
-//                 HeightTextureState::Available => {
-//                     // draw
-//                     if let Some(height_texture) = height_texture {
-//                         mesh.draw(render_pass, &height_texture, &texture, instance);
-//                     }
-//                 }
-//             }
-//         }
-//     }
-// }
-
-// struct TextureLod {
-//     texture: [Option<deferred_heightmap_shader::Texture>; LOD],
-
-//     position: cgmath::Vector3<f32>,
-// }
-
-// impl TextureLod {
-//     fn get(
-//         &self,
-//         view_position: &cgmath::Vector3<f32>,
-//     ) -> &Option<deferred_heightmap_shader::Texture> {
-//         let distance = view_position.distance(self.position);
-
-//         return &self.texture[0];
-//     }
-// }
-
-// enum HeightTextureState {
-//     NotRequested,
-//     IsRequested,
-//     Available,
-// }
+        self.requests = quad_tree_draw.requests.clone();
+    }
+}
