+++ market_economy_simulation/src/terrain_storage/quad_tree_draw.rs
@@ -1,3 +1,6 @@
+//! Object to traverse the lod quad tree and draw every leaf
+//!
+
 use std::usize;
 
 use super::{
