+++ market_economy_simulation/src/terrain_storage/terrain_texture_details.rs
@@ -1,6 +1,4 @@
-
-
-
+//! Info for terrain data generation
 
 #[derive(Clone, Debug)]
 pub struct TerrainTextureDetails {
